% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_window.R
\name{apply_window_function}
\alias{apply_window_function}
\title{Apply window function}
\usage{
apply_window_function(
  df,
  col = "combined",
  fun = c("threshold", "binary"),
  w = 1000,
  t = 0.5
)
}
\arguments{
\item{df}{a dataframe}

\item{col}{the column of the dataframe to apply function to}

\item{fun}{the filter function to apply to the window}

\item{w}{the window size}

\item{t}{the threshold (only used of fun = "threshold")}
}
\value{
a dataframe with added filtered column
}
\description{
Applies a window function to the supplied dataframe. Available window functions are threshold and binary.
}
