% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_window.R
\name{combine_to_accvector}
\alias{combine_to_accvector}
\title{Combine acceleration vector}
\usage{
combine_to_accvector(sourcefile)
}
\arguments{
\item{sourcefile}{path to .csv file with time and three dimensions (x, y, and z)}
}
\value{
dataframe with added combined vector
}
\description{
Helper function to import .csv, convert time to POSIXct and create a combined vector.
}
