% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_filter.R
\name{convert_grayscale}
\alias{convert_grayscale}
\title{Convert to grayscale}
\usage{
convert_grayscale(image)
}
\arguments{
\item{image}{a magick image}
}
\value{
a magick image
}
\description{
Legacy, instead use convert_rgb which builds further on this function allowing conversion to grayscale or cutoff
}
