% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_search.R
\name{count_matches}
\alias{count_matches}
\title{Count matches in words using pattern list}
\usage{
count_matches(words, patterns)
}
\arguments{
\item{words}{character vector of words}

\item{patterns}{character vector of regex patterns}
}
\value{
integer count
}
\description{
Count matches in words using pattern list
}
