% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_customer.R
\name{get_best_customer}
\alias{get_best_customer}
\title{get_best_customer}
\usage{
get_best_customer(x, date)
}
\arguments{
\item{x}{a data object with columns TransDate, Customer, PurchAmount, Cost}

\item{date}{a specification for the time of interest. Must be of format "yyyy-mm-dd"}
}
\value{
The Customer with the highest profit for a given date \code{time} of data \code{x}
}
\description{
Function that calculates the customer with the best monthly return for a given date.
}
\details{
\code{data} contains the transaction data. The data set must contain a
            column labeled "Customer" that allows unique customer identification
            and a column labeled "TransDate", indicating the purchase date.
            The column "PurchAmount" specifies the total spending per purchase.
}
\examples{
get_best_customer(transactions, date="2008-10-01")
get_best_customer(transactions, date="2008-01-01")

}
