% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_filter.R
\name{iterative_converter}
\alias{iterative_converter}
\title{Iterative converter}
\usage{
iterative_converter(
  image,
  conversion = c("grayscale", "red", "green", "blue", "cutoff"),
  percent = 50
)
}
\arguments{
\item{image}{an image}

\item{conversion}{conversion to apply}

\item{percent}{percentage to use for cutoff}
}
\value{
An image with applied conversions
}
\description{
This function is just for demonstration purposes and according to the instructions for the exam.
I highly recommend not to use it.
This per pixel iterative approach is O(^2) and slow. A vectorized approach is preferred.
For a small image this is 100 times slower than the vectorized approach.
}
