% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_window.R
\name{make_plot}
\alias{make_plot}
\title{Create plot}
\usage{
make_plot(
  sourcefile,
  type = c("threshold", "binary"),
  windowsize = 1000,
  threshold = 0.5
)
}
\arguments{
\item{sourcefile}{a .csv file with at least time, and three dimensions x, y, and z}

\item{type}{type of window function to apply, mean or binary (default = mean)}

\item{windowsize}{size of window (default = 1000)}

\item{threshold}{threshold to apply (default = 0.5)}
}
\value{
a plot
}
\description{
Reads a source .csv containing at least time, x, y and z.
Makes a combined acceleration vector using the eucledian method.
Applies a sliding window filter to the plot using either mean or binary method
}
\examples{
# Loading example data, you shoud use your own .csv file
csv_file <- system.file("extdata", "testwindow.csv", package = "exampackage")
# Plot with base settings
make_plot(csv_file)
# Threshold plot with adjusted window and threshold
make_plot(csv_file, type = "threshold", windowsize = 500, threshold = 1)
# Binary plot
make_plot(csv_file, type = "binary", windowsize = 200)

}
