% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_filter.R
\name{modify_pixel}
\alias{modify_pixel}
\title{Modify pixel}
\usage{
modify_pixel(
  pixel,
  conversion = c("grayscale", "red", "green", "blue", "cutoff"),
  percent = 50
)
}
\arguments{
\item{pixel}{a raw pixel}

\item{conversion}{conversion to apply}

\item{percent}{percentage to use for cutoff}
}
\value{
A raw pixel with applied conversion
}
\description{
This function is just for demonstration purposes and according to the instructions for the exam.
This per pixel iterative approach is O(^2) and slow. A vectorized approach is preferred.
}
