% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_search.R
\docType{data}
\name{neg_patterns}
\alias{neg_patterns}
\title{Negative word patterns}
\format{
An object of class \code{character} of length 27.
}
\usage{
neg_patterns
}
\description{
Negative word patterns
}
\keyword{internal}
