% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_search.R
\docType{data}
\name{pos_patterns}
\alias{pos_patterns}
\title{Positive word patterns}
\format{
An object of class \code{character} of length 20.
}
\usage{
pos_patterns
}
\description{
Positive word patterns
}
\keyword{internal}
