% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_search.R
\name{text_mood}
\alias{text_mood}
\title{Text mood}
\usage{
text_mood(text)
}
\arguments{
\item{text}{Path to a text file}
}
\value{
character string: "Text is positive :)", "neutral :|", or "negative :("
}
\description{
Text mood
}
