% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\docType{data}
\name{transactions}
\alias{transactions}
\title{Transactions data}
\format{
An object of class \code{data.frame} with 223191 rows and 7 columns.
}
\usage{
data(transactions)
}
\description{
Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.
}
\examples{
data(transactions)
summary(transactions)
}
\keyword{datasets}
