% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_window.R
\name{window_binary}
\alias{window_binary}
\title{Window binary}
\usage{
window_binary(vec)
}
\arguments{
\item{vec}{a vector from a window}
}
\value{
+1 or -1 depending of mean change in window
}
\description{
Window binary
}
