% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_window.R
\name{window_threshold}
\alias{window_threshold}
\title{Window threshold}
\usage{
window_threshold(vec, t)
}
\arguments{
\item{vec}{a vector from a window}

\item{t}{threshold value}
}
\value{
new value if change exceed threshold
}
\description{
Window threshold
}
