% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_filter.R
\name{convert_rgb}
\alias{convert_rgb}
\title{Convert RGB}
\usage{
convert_rgb(
  image,
  conversion = c("grayscale", "red", "green", "blue", "cutoff"),
  percent = 50
)
}
\arguments{
\item{image}{A magick image or a path to an image}

\item{conversion}{Conversion to apply, grayscale, red, green, blue, or cutoff}

\item{percent}{Percentage to use for cutoff}
}
\value{
An image with applied conversions
}
\description{
Convert RGB
}
