% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_filter.R
\name{image_to_cutoff}
\alias{image_to_cutoff}
\title{Image to cutoff}
\usage{
image_to_cutoff(image_path, output = "default", cutoff = 50)
}
\arguments{
\item{image_path}{path to an image file}

\item{output}{path to output file, defaults to input with added _grayscale}

\item{cutoff}{a percentage value where to apply cutoff}
}
\value{
image with cutoff applied
}
\description{
Image to cutoff
}
\examples{
# Load test image, you should use your own image
image_file <- system.file("extdata", "testimage.jpg", package = "exampackage")
# Temporary file for output
output_file <- tempfile(fileext = ".jpg")
# Apply cutoff function
image_to_cutoff(image_file,output_file)
# Apply cutoff function with custom cutoff
#' image_to_cutoff(image_file, output_file, cutoff = 30)

}
