% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_filter.R
\name{image_to_grayscale}
\alias{image_to_grayscale}
\title{Image to Grayscale}
\usage{
image_to_grayscale(image_path, output = "default")
}
\arguments{
\item{image_path}{path to an image file}

\item{output}{path to output file, defaults to input with added _grayscale}
}
\value{
a grayscale iamge
}
\description{
Converts an image to grayscale
}
