% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_search.R
\name{text_mood}
\alias{text_mood}
\title{Text mood}
\usage{
text_mood(text)
}
\arguments{
\item{text}{Path to a text file}
}
\value{
Prints summary. Returns a list of ratio, positive and negative word counts.
}
\description{
Detects the mood of the text
}
\examples{
# Load test file
text_file <- system.file("extdata", "testtext.txt", package = "exampackage")
# Run function
text_mood(text_file)

}
